Refactor: Move UI Base fragments to ui/base subpackage
Moved `BaseDashboardFragment`, `BaseTeamFragment`, and `BaseVoicesFragment` to `org.ole.planet.myplanet.ui.base` package.
Updated package declarations and imports in all referencing files.
This change organizes the base UI components into a dedicated package for better structure.
