+++ app/src/main/java/org/ole/planet/myplanet/ui/community/CommunityServicesFragment.kt
@@ -11,11 +11,11 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.databinding.FragmentCommunityServicesBinding
 import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmNews
-import org.ole.planet.myplanet.ui.teams.BaseTeamFragment
 import org.ole.planet.myplanet.ui.teams.TeamDetailFragment
 import org.ole.planet.myplanet.utilities.MarkdownUtils.prependBaseUrlToImages
 import org.ole.planet.myplanet.utilities.MarkdownUtils.setMarkdownText
 import org.ole.planet.myplanet.utilities.NavigationHelper.replaceFragment
+import org.ole.planet.myplanet.ui.base.BaseTeamFragment
 
 class CommunityServicesFragment : BaseTeamFragment() {
     private var binding: FragmentCommunityServicesBinding? = null
